# server/swagger.yaml
openapi: 3.0.3
info:
  title: Bike Rental API
  version: "1.0.0"
  description: |-
    Simple bike rental backend (Express + MySQL using mysql2) — OpenAPI specification generated from the server source.
servers:
  - url: http://localhost:4000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: alice
        role:
          type: string
          example: user
        email:
          type: string
          format: email
        phone:
          type: string
        age:
          type: integer
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
...
paths:
  /auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                email:
                  type: string
                  format: email
                phone:
                  type: string
                age:
                  type: integer
              required: [username, password]
      responses:
        "200":
          description: Created user
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  username:
                    type: string
  /auth/login:
    post:
      summary: Login and receive a JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required: [username, password]
      responses:
        "200":
          description: Authentication result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /profile:
    get:
      summary: Get current user's profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      summary: Update current user's profile
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
                email: { type: string }
                phone: { type: string }
                age: { type: integer }
      responses:
        "200": { description: updated user basic info }
...